Improved FindWinners() and FindNeighMaxLocalError()
In FindNeighMaxLocalError() there were no check whether the neighbour was found, because sometimes the error value doesn't change and still equals 0. By improving this function we can be sure that there is at least one neighbour found so that we can add new node between this neighbour and parent node.